Refactor request generation and auth
Basically this is to deal with the fact that there are parts of the URL
that we don't necessarily know before logging in, which makes the
on-demand thing a little hard. Deal with it more elegantly.

Also, use introspection to allow both form-data POSTs and JSON posts.
url.Values will be sent as an application/x-www-form-urlencoded,
everything else will be json.Marshal()ed.